backend/requirements.txt
# FleetAI Backend Dependencies

# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
starlette>=0.36.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.25
aioodbc>=0.5.0
pyodbc>=5.1.0
aiosqlite>=0.19.0
alembic>=1.13.0

# Pydantic
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0

# Azure OpenAI (optional)
openai>=1.12.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Async Support
aiofiles>=23.2.0

# Logging & Monitoring
structlog>=24.1.0
python-json-logger>=2.0.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
croniter>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
